feat: 진영별 추가 점수 구현

+++ src/main/java/view/dto/ScoreDto.java
@@ -4,6 +4,6 @@ import domain.game.Side;
 
 public record ScoreDto(
         Side side,
-        int score
+        double score
 ) {
 }
